+++ eox_hooks/edxapp_wrapper/backends/enrollments_j_v1.py
@@ -1,5 +1,5 @@
 """
-Backend file for Course Enrollment.
+Backend file for Course Enrollment related objects.
 """