+++ eox_hooks/settings/test.py
@@ -29,6 +29,10 @@ EOX_HOOKS_TEST_ENVIRONMENT = True
 
 EOX_HOOKS_ENROLLMENTS_BACKEND = "eox_hooks.edxapp_wrapper.backends.enrollments_j_v1"
 
+EOX_HOOKS_COURSES_BACKEND = "eox_hooks.edxapp_wrapper.backends.courses_j_v1"
+
+EOX_HOOKS_COURSE_MODES_BACKEND = "eox_hooks.edxapp_wrapper.backends.course_modes_j_v1"
+
 # This key needs to be defined so that the check_apps_ready passes and the
 # AppRegistry is loaded
 DATABASES = {